requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
et_xmlfile==2.0.0
idna==3.10
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
